pyproject.toml
[project]
name = "postgres-json"
version = "0.1.0"
description = "Learn how to work with JSON data in PostgreSQL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython-sql>=0.5.0",
    "jupysql>=0.11.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.45",
]
